% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_profile.R
\name{clean_profile}
\alias{clean_profile}
\title{Clean up a profile}
\usage{
clean_profile(data, remove_lower = 0.5)
}
\arguments{
\item{data}{The profile data (`data.frame``) containing a column named either "depth" or "press"
and a column named "dt" containing the time stamp of the measurement}

\item{remove_lower}{Numeric value of the thickness of the bottom layer to truncate}
}
\value{
The cleaned profile data
}
\description{
Strips unwanted values from a profile, usually read from raw data.
}
\details{
This function removes rows in a profile that:
\itemize{
\item have depth less than zero (ie in air),
\item were taken when the sonde was pulled back up
\item are within a certain distance of the maximum depth, given by \code{remove_lower}.
The idea of \code{remove_lower} is to remove the lowest values when the sonde hits the sediment.
}
}
\examples{
\dontrun{
pr_loc <- system.file("extdata", package="seefo")
pr_name <- file.path(pr_loc,  "profiles","20170801","YT1_20170801_1.CTM644")
pr1 <- read_ctm(pr_name, clean=FALSE)
tail(pr1)

clean_profile(pr1)
tail(pr1)
}

}
\seealso{
\code{\link{get_profile_info}}
}
\author{
Tom Shatwell
}
