% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_rho.R
\name{delta_rho}
\alias{delta_rho}
\title{Fast calculation of mixed layer depth (depth of density threshold)}
\usage{
delta_rho(T, z, drho = 0.1)
}
\arguments{
\item{T}{Temperature data. Must be a mxn matrix with n temperature profiles in the columns}

\item{z}{The depths corresponding to the temperature profiles. Numeric vector of length m.}

\item{drho}{The density threshold threshold depths corresponding to the temperature profiles. Numeric vector of length m.}
}
\value{
A vector of mixed layer depths
}
\description{
Fast calculation of the depth of a set density threshold
}
\details{
Calculates mixed layer depths as the depth of when the given density gradient \code{drho} is first exceeded.
Designed to be as fast as possible for large data sets. Does very minimal checks to remove \code{Inf}s and \code{NaN}s, so be careful.
Choose the right vertical resolution in the temperatures to ensure a more robust result.
See Wilson et al. 2020, HESS XXXX
}
\examples{
\dontrun{
z <- c(0,2,4,6,10)
T <-  matrix(c(15,14,10,8,7,
16,13,12,11,11,
22,19,18,14,13),
nrow=length(z))
delta_rho(T,z)
}

}
\seealso{
\code{\link{h2mix}},  \code{\link{delta_rho}}
}
\author{
Tom Shatwell
}
