% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5cli.R
\name{era5cli}
\alias{era5cli}
\title{Generate a command to download era5 data with the era5cli tool}
\usage{
era5cli(vars, startyear, endyear, prefix, target, coords)
}
\arguments{
\item{vars}{Character vector containing names of ERA5 variables to download}

\item{startyear}{Numeric length 1}

\item{endyear}{Numeric length 1}

\item{prefix}{A label to begin the downloaded filenames with}

\item{target}{Not used}

\item{coords}{Vector (length 2) containing latitude and longitude}
}
\value{
A character vector length 1 with the command and arguments to use with the era5cli tool
}
\description{
Generates a character string you can paste into a shell
}
\details{
The location and variable name are kept in the attributes of the output.
}
\examples{
\dontrun{
vars <- c(
"mean_surface_downward_long_wave_radiation_flux", # W m**-2
"mean_surface_downward_short_wave_radiation_flux", # W m**-2
"total_cloud_cover", # 0..1
"2m_temperature", # K
"surface_pressure", # Pa
"2m_dewpoint_temperature", # K
"10m_u_component_of_wind", # m s**-1
"10m_v_component_of_wind" # m s**-1
)
 era5cli(vars=vars,startyear=2013,endyear=2019,
 prefix = "era_kilpisjarvi",
 target = "../nc_input/era5/",
 coords = c(69.03, 20.8))
}

}
\author{
Tom Shatwell
}
