% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinction_coeff.R
\name{extinction_coeff}
\alias{extinction_coeff}
\title{Calculate the extinction coefficient from PAR}
\usage{
extinction_coeff(PAR, z, plot_profile = FALSE)
}
\arguments{
\item{PAR}{The PAR measurements (numeric)}

\item{z}{Depths of PAR measurements (numeric)}

\item{plot_profile}{Should the data be plotted, showing the fitted extinction coefficient (logical)? Only for profiles with more than 2 measurements.}
}
\value{
A numeric length 1.
}
\description{
Calculates the light extinction (attenuation) coefficient from measurements of PAR data according to the Lambert-Beer law
}
\details{
Extinction is best calculated from pairs of simultaneous PAR measurements at different depths with 2
different sensors. The function will assume this is the case if the \code{length(PAR)} and \code{length(z)} = 2.
In this case the function will calculate the extinction analytically from the Lambert-Beer law.
Sometimes only a (non-simultaneous) profile of PAR measurmentns is available, and the function will assume this is the case if \code{length(PAR)} and \code{length(z)} > 2.
In this case it will calculate the extinction coefficient as the slope of the regression through the logged PAR measurements.
}
\examples{

\dontrun{
# pair of measurements with 2 sensors

depth <- c(0.75, 1.25)
light <- c(1263, 957)

extinction_coeff(PAR=light, z=depth)

# with a profile of measurements

depth <- c(0.5, 1, 1.5, 2, 2.5, 3)
light <- c(769, 532, 370, 211, 158, 95)

extinction_coeff(light,depth, plot_profile=TRUE)

}

}
\author{
Tom Shatwell
}
