% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile_info.R
\name{get_profile_info}
\alias{get_profile_info}
\title{Extracts information from a sonde filename}
\usage{
get_profile_info(filename)
}
\arguments{
\item{filename}{Filename to examine (character), including the file extension.}
}
\value{
A character vector with elements containing the location, date, profile number, file extension, and possibly comments.
Returns \code{NA} if any of these are missing.
}
\description{
Extracts information from a sonde raw data filename, like the measuring location, date, and sonde number.
}
\details{
This function parses the filename and extracts information according to the limnophsycs naming convention.
It assumes that the filename has the structure \code{"LOC_DATE_PROFILENUMBER_OPTIONALSTUFF.SONDENUM"}.
It separates the name into chunks separated by \code{"_"}. Whole paths can be given too. In this case the path discarded from the filename.
}
\examples{
\dontrun{
get_profile_info("../mydata/lakes/YH1_20210101_1.CTM644")
}

}
\author{
Tom Shatwell
}
