% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2mix.R
\name{h2mix}
\alias{h2mix}
\title{Fast calculation of mixed layer depth (lower thermocline boundary)}
\usage{
h2mix(T, z)
}
\arguments{
\item{T}{Temperature data. Must be a mxn matrix with n temperature profiles in the columns}

\item{z}{The depths corresponding to the temperature profiles. Numeric vector of length m.}
}
\value{
A vector of mixed layer depths
}
\description{
Fast calculation of lower thermocline boundary
}
\details{
Calculates mixed layer depths as the depth of the maximum curvature of the temperature profile (d2T/dz2).
Designed to be as fast as possible for large data sets, so doesn't do any checks at all (be careful).
Choose the right vertical resolution in the temperatures to ensure a more robust result.
}
\examples{
\dontrun{
z <- c(0,2,4,6,10)
T <-  matrix(c(15,14,10,8,7,
16,13,12,11,11,
22,19,18,14,13),
nrow=length(z))
h2mix(T,z)
}

}
\seealso{
\code{\link{hmix}},  \code{\link{h3mix}}, \code{\link{delta_rho}}
}
\author{
Georgiy Kirillin and Tom Shatwell
}
