% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3mix.R
\name{h3mix}
\alias{h3mix}
\title{Robust calculation of mixed layer depth}
\usage{
h3mix(
  T,
  z,
  plot = FALSE,
  thresh = 1,
  therm.dz = 1,
  min.hmix = 1.5,
  var = "hmix",
  min.gradient = 0.3,
  ...
)
}
\arguments{
\item{T}{Temperature profile as numeric vector.}

\item{z}{The depths corresponding to the temperature profile, as a numeric vector.}

\item{plot}{Should a plot of the profile showing the estimates be shown? (logical).}

\item{thresh}{The threshold for the difference between top and bottom temperature to infer stratification.}

\item{therm.dz}{Representative thickness of the thermocline layer. 1 m works well from experience.}

\item{min.hmix}{I can't really remember what this is for.}

\item{var}{Which metric should be returned? One of \code{"hmix"}, \code{"htherm"}, \code{"both"} for mixed layer depth, thermocline depth, or both, respectively.}

\item{min.gradient}{Not used.}

\item{...}{Arguments passed to \code{plot()}, when \code{plot=TRUE}.}
}
\value{
A vector of length 2 with the mixed layer depth \code{hmix} and the thermocline depth \code{htherm}
}
\description{
Robust calculation of the mixed layer depth and thermocline depth,
more suited for manual or irregular
temperature data.
}
\details{
This function estimates the mixed depth and the thermocline depth.
It uses regressions to locate the depth of the kink in the temperature profile
as the mixed layer depth, and the maximum gradient as the thermocline.
It is better suited to irregular, esp manually measured profile data.
It first finds the minimum curvature in the
profile (maximum for winter stratification) as the initial guess of the
border of the surface mixed
layer. Then it finds the thermocline as the depth of maximum T-gradient. It
estimates the mixed layer depth as the depth where the regression line through
the surface layer temperatures intersects with the regression line through the
thermocline temperatures, extended by a window of thickness \code{therm.dz}. It performs some other checks, like whether
stratification exists according to whether the surface-bottom temperature difference > \code{thresh},
and whether the thermocline
is above the surface layer, or whether the lake is completely isothermal and
there is no intersection (and returns \code{NA}). If mixed, it assumes the mixed
layer depth is the maximum depth. It also plots the profile if desired
(\code{plot=TRUE}), marking the surface layer and inflection point in red and the
thermocline in blue.
Surface temperature is estimated as the mean of the upper 2 m, or the top two measurements,
bottom temperature is estimated as the mean of the deepest 2 m layer.
}
\examples{
\dontrun{
T <- c(25.44,25.46,25.48,25.46,25.44,24.36,20.48,18.09,15.96,13.95,11.67,
10.62,9.82,9.51,9.22,9.01,8.81,8.68,8.58,8.39,8.29,8.14,8.08,8.03,8.01,
8.05,7.94,7.93,7.85)
z <- 0:28
h3mix(T,z,plot=TRUE)
}

}
\seealso{
\code{\link{hmix}}, \code{\link{h2mix}},  \code{\link{delta_rho}}
}
\author{
Tom Shatwell
}
