% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmix.R
\name{hmix}
\alias{hmix}
\title{Fast calculation of mixed layer depth (upper thermocline boundary)}
\usage{
hmix(T, z)
}
\arguments{
\item{T}{Temperature data. Must be a mxn matrix with n temperature profiles in the columns}

\item{z}{The depths corresponding to the temperature profiles. Numeric vector of length m.}
}
\value{
A vector of mixed layer depths
}
\description{
Fast calculation of the upper thermocline boundary
}
\details{
Calculates mixed layer depths as the depth of the maximum temperature gradient (dT/dz).
Designed to be as fast as possible for large data sets. It doesn't do any checks at all,
and only works with regular data, so be careful.
Choose the right vertical resolution in the temperatures to ensure a more robust result.
Use the rLakeAnalyzer functions for more robust solutions which are
orders of magnitude slower.
}
\examples{
\dontrun{
z <- c(0,2,4,6,10)
T <-  matrix(c(15,14,10,8,7,
16,13,12,11,11,
22,19,18,14,13),
nrow=length(z))
hmix(T,z)
}

}
\seealso{
\code{\link{h2mix}},  \code{\link{delta_rho}}
}
\author{
Georgiy Kirillin and Tom Shatwell
}
