% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap.R
\name{leap}
\alias{leap}
\title{Calculation of leap years}
\usage{
leap(yr)
}
\arguments{
\item{yr}{Year (numeric)}
}
\value{
Returns a vector the same length as \code{yr} containing \code{0} for non-leap years, and \code{1} for leap years.
}
\description{
Determines whether a year is a leap year
}
\examples{
\dontrun{
leap(c(2000,2002,2004,2006,2008,2010))
}

}
\author{
Tom Shatwell
}
