% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ctm.R
\name{read_ctm}
\alias{read_ctm}
\title{Reads profile data from raw output file from a CTM CTD}
\usage{
read_ctm(filename, keep = "all", header = "auto", TZ = "UTC", clean = TRUE)
}
\arguments{
\item{filename}{Name of the file to read, including the file extension}

\item{keep}{names of the columns to retain in the output. Defaults to `"all"``}

\item{header}{the names of the columns in the raw data file. Defaults to \code{"auto"}}

\item{TZ}{time zone given to the output datetimes, defaults to `"UTC"`` (recommended)}

\item{clean}{Should the profile be cleaned? (logical)}
}
\value{
A \code{data.frame} containing the profile data
}
\description{
Reads a data from an output file generated by a CTM CTD.
}
\details{
This function first gleans information from the filename using \code{get_profile_info()},
including the sampling location from the beginning of the filename,
and the sonde number from the file extension.
It parses the file and begins to read data 2 lines after the keyword "Datasets".
It assigns the names in \code{header} to the columns. If \code{header} is set to \code{"auto"},
it will assign hard coded names according to the sonde number, hoping that one sonde always thas the same sensors attached.
It converts the date and time columns to a POSIX datetime with the timezone \code{TZ}.
The function finally drops any columns that are not named in \code{keep}.
If \code{keep} contains names not found in the header, it will add a column of \code{NA}s for each unfound name.
The \code{header} can include \code{c("no","press","temp","cond","cond_hi","salin","DO_sat","DO","pH","turb","chla","BGAPE","BGAPC","sv","intD","intT")}
}
\examples{
\dontrun{
pr_loc <- system.file("extdata", package="seefo")
pr_name <- file.path(pr_loc,  "profiles","20170801","YT1_20170801_1.CTM644")
pr1 <- read_ctm(pr_name)
head(pr1)
}

}
\seealso{
\code{\link{get_profile_info}}, \code{\link{read_bbe}}, \code{\link{clean_profile}}, \code{\link{batch_read}}
}
\author{
Tom Shatwell
}
