% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ret_eff_data}
\alias{ret_eff_data}
\title{Inflow and outflow concentration time series}
\format{
A data frame with 5004 rows and 7 variables:
\describe{
\item{date}{date in POSIXct format}
\item{year}{Year corresponding to date}
\item{doy}{Day of year}
\item{tributary}{Name of the tributary}
\item{variable}{Name nutrient species}
\item{value}{Measured concentration}
\item{in_outlet}{Whether the tributary is an inflow or outflow}
...
}
}
\source{
Taynara Fernadez
}
\usage{
ret_eff_data
}
\description{
A dataset containing the inflow and outflow nutrient concentrations
}
\keyword{datasets}
