% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention_eff.R
\name{retention_eff}
\alias{retention_eff}
\title{Nutrients load and retention efficiency calculator}
\usage{
retention_eff(data, methods, start.year, end.year)
}
\arguments{
\item{data}{Must be a dataframe consisting of date, variable, value, in_outlet and tributary}

\item{methods}{Vector containing one or more of the three available methods ("GAM.load", "method1", "method2")}

\item{start.year}{Numeric length 1}

\item{end.year}{Numeric length 1}
}
\value{
A dataframe with annual load and retention efficiency for each nutrient
}
\description{
Calculates the nutrient load in tons/year using Standard Method 1 and 2 according to Hilde 2003 and Generalized Additive Model (GAM).
}
\details{
Input data frame must consist of:
date must be mm/dd/yyyy format;
variable is a character vector with nutrient's name;
value is a numeric vector with the observed concentrations in mg/l and the discharge in m3/s;
in_outlet is a character vector naming "inflow" or "outflow";
tributary is a character vector naming the pre-dams.
}
\examples{
#Load it as binary data for the example
\dontrun{
data <- read.table("data/retention_eff.csv", header=T, sep=",", dec=".") #.rda
methods <- c("method1","method2", "GAM.load")
start.year=2000
end.year=2017
}

}
\author{
Karsten Rinke, Taynara Fernandes and Tom Shatwell
}
