% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_water.R
\name{rho_water}
\alias{rho_water}
\title{Calculates density of freshwater}
\usage{
rho_water(t)
}
\arguments{
\item{t}{Temperature in degrees Celsius (numeric)}
}
\value{
numeric vector the same length as \code{t}
}
\description{
Calculates the density of fresh water.
}
\details{
Calculates the density from temperature (kg m^-3) using the formula of Millero & Poisson (1981).
this is the method stated in the isimip 2b protocol in July 2019
}
\examples{
\dontrun{
rho_water(0:25)
}

}
\seealso{
\code{\link{hmix}},  \code{\link{delta_rho}}
}
\author{
Georgiy Kirillin and Tom Shatwell
}
