% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seefo.R
\name{seefo}
\alias{seefo}
\title{A collection of useful tools for lake research.}
\usage{
seefo()
}
\description{
These tools include: calculating ice and stratification phenology (\code{\link{strat_phenology}}, \code{\link{ice_phenology}}),
reading raw data files from sondes (\code{\link{batch_read}},  \code{\link{read_bbe}},  \code{\link{read_ctm}}, \code{\link{read_ixx}}),
calculating water density \code{\link{delta_rho}},
calculating volume-weighted mean using hyspographic information (to be added),
interpolating to fixed depth grids (to be added) or daily values (to be added),
calculating the mixed layer depth (\code{\link{hmix}}, \code{\link{h2mix}}, \code{\link{h3mix}}).
}
