% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar2PAR.R
\name{solar2PAR}
\alias{solar2PAR}
\title{Convert solar radiation into PAR}
\usage{
solar2PAR(solar, parfrac = 0.5, reflect = 0.1, joule2umol = 4.56)
}
\arguments{
\item{solar}{Solar radiation in W m^-2 (numeric)}

\item{parfrac}{The fraction of solar radiation that is PAR (400 - 700 nm) (numeric)}

\item{reflect}{The loss due to reflection and backscatter at the surface of the water (numeric)}

\item{joule2umol}{Conversion factor for solar radiation to PAR}
}
\value{
A numeric vector the same length as \code{solar}.
}
\description{
Converts solar (shortwave) radiation in W/m2 above the water surface into PAR (photosynthetically active radiation in umol / m2 / s) just below the water surface
}
\details{
This function helps to convert solar radiation usually available from weather stations into PAR.
Solar radiation should be in W/m2. The assumptions are that 50\% of solar radiation is PAR (400 - 700 nm),
and 10 \% of that is lost at the water surface due to reflection.
The conversion to micromoles of photons (400-700) depends on the spectrum but the default value is typical for sunlight.
}
\examples{

\dontrun{
solar2PAR(c(100, 200, 300))
}

}
\author{
Tom Shatwell
}
