% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ts_Tb_ice}
\alias{Ts_Tb_ice}
\title{Surface and bottom temperatures and ice thickness}
\format{
A data frame with 3653 rows and 4 variables:
\describe{
\item{date}{date in POSIXct format}
\item{Ts}{The surface temperature in degrees Celsius}
\item{Tb}{The bottom temperature in degrees Celsius, depth ~23 m}
\item{H_ice}{The ice thickness in meters}
...
}
}
\source{
Tom Shatwell
}
\usage{
Ts_Tb_ice
}
\description{
A dataset containing the surface and bottom temperatures as well as ice thickness from
historical model simulations of Lake Stechlin with the model FLake version 2.
}
\keyword{datasets}
