% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_read.R
\name{batch_read}
\alias{batch_read}
\title{Reads raw sonde profile data from several files and compiles the results}
\usage{
batch_read(
  sonde,
  path,
  fnames = "autosearch",
  keep = "all",
  recursive = TRUE,
  clean = TRUE,
  TZ = "UTC"
)
}
\arguments{
\item{sonde}{Type of sonde. One of \code{"ctm"},\code{"bbe"},\code{"ixx"}.}

\item{path}{The path to a directory in which to search for matching files. Used only if \code{fnames}== \code{"autosearch"}}

\item{fnames}{The names of the files to read. Defaults to \code{"autosearch"},
in which case all filenames with an extension matching the sonde type will be read.
If \code{fnames} != \code{"autosearch"}, then \code{path} and \code{recursive} are not used.}

\item{keep}{names of the columns to retain in the output. Defaults to \code{"all"}}

\item{recursive}{Should subdirectories in \code{path} also be searched? (logical, only if \code{fnames}== \code{"autosearch"})}

\item{clean}{Should the profiles be cleaned, by removing measurements in air, near the sediment, and on the upward pull?? (logical)}

\item{TZ}{time zone given to the output datetimes, defaults to \code{"UTC"} (recommended)}
}
\value{
A \code{data.frame} containing the profile data
}
\description{
Reads raw sonde profile data from several files and compiles them into a data.frame.
By default, it searches a given directory and all subdirectories for matching files for a given sonde type,
cleans up the data, and returns it sorted and labelled.
}
\details{
To be added later.
}
\examples{
\dontrun{
out <- batch_read(sonde="ctm", path="../data/")
}

}
\seealso{
\code{\link{get_profile_info}},  \code{\link{read_bbe}},  \code{\link{read_ctm}}
}
\author{
Tom Shatwell
}
