% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5cli.R
\name{era5cli}
\alias{era5cli}
\title{Generate a command to download era5 data with the era5cli tool}
\usage{
era5cli(vars, startyear, endyear, prefix, target, coords)
}
\arguments{
\item{vars}{Character vector containing names of ERA5 variables to download. See ERA5 longname in Details.}

\item{startyear}{Numeric length 1}

\item{endyear}{Numeric length 1}

\item{prefix}{A label to begin the downloaded filenames with}

\item{target}{Not used}

\item{coords}{Vector (length 2) containing latitude and longitude}
}
\value{
A character vector length 1 with the command and arguments to use with the era5cli tool
}
\description{
Generates a character string you can paste into a shell
}
\details{
The location and variable name are kept in the attributes of the output.\tabular{lll}{
   \strong{ERA5 long name} \tab \strong{ERA5 shortname} \tab \strong{Units} \cr
   surface_solar_radiation_downwards \tab ssrd \tab J m**-2 \cr
   surface_solar_radiation_downward_clear_sky \tab ssrdc \tab J m**-2 \cr
   mean_surface_downward_short_wave_radiation_flux \tab msdwswrf \tab W m**-2 \cr
   mean_surface_downward_short_wave_radiation_flux_clear_sky \tab msdwswrfcs \tab W m**-2 \cr
   mean_surface_downward_long_wave_radiation_flux \tab msdwlwrf \tab W m**-2 \cr
   surface_thermal_radiation_downwards \tab strd \tab J/m^2 \cr
   2m_temperature \tab t2m \tab K \cr
   surface_pressure \tab ?? \tab Pa \cr
   2m_dewpoint_temperature \tab d2m \tab K \cr
   10m_u_component_of_wind \tab u10 \tab m/s \cr
   10m_v_component_of_wind \tab v10 \tab m/s \cr
   mean_total_precipitation_rate \tab mtpr \tab kg m**-2 s**-1 \cr
   total_precipitation \tab ?? \tab m \cr
   total_cloud_cover \tab tcc \tab 0..1 \cr
   low_cloud_cover \tab lcc \tab 0..1 \cr
   medium_cloud_cover \tab mcc \tab 0..1 \cr
   high_cloud_cover \tab hcc \tab 0..1 \cr
   cloud_base_height \tab cbh \tab m \cr
}
}
\examples{
\dontrun{
vars <- c(
"mean_surface_downward_long_wave_radiation_flux", # W m**-2
"mean_surface_downward_short_wave_radiation_flux", # W m**-2
"total_cloud_cover", # 0..1
"2m_temperature", # K
"surface_pressure", # Pa
"2m_dewpoint_temperature", # K
"10m_u_component_of_wind", # m s**-1
"10m_v_component_of_wind" # m s**-1
)
 era5cli(vars=vars,startyear=2013,endyear=2019,
 prefix = "era_kilpisjarvi",
 target = "../nc_input/era5/",
 coords = c(69.03, 20.8))
}

}
\author{
Tom Shatwell
}
