% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_era5_var.R
\name{gather_era5_var}
\alias{gather_era5_var}
\title{Import and collate ERA5 variable from different files at a location}
\usage{
gather_era5_var(
  namestrings = c(".nc$"),
  varid,
  lon,
  lat,
  datapath,
  files = "autosearch",
  recursive = FALSE,
  radius = c(0.5, 1)
)
}
\arguments{
\item{namestrings}{list of vector containing text fragments to uniquely match the  filenames. Typically it will contain the file extension (".nc$"), variable name (e.g. "2m_temperature", and perhaps a prefix (e.g. "era5"). This is only used if files is "autosearch".}

\item{varid}{The name of the variable in the nc file.}

\item{lon}{Longitude (numerical) of the location to collate data.}

\item{lat}{Latitude (numerical) of the location to collate data.}

\item{datapath}{The path to the folder in which to search for the files (character)}

\item{files}{The filenames to be read (character vector). The default is \code{"autosearch"}, in which case the files are matched with the patterns in namestrings.}

\item{recursive}{Should subfolders also be searched? (logical)}

\item{radius}{The radius (numerical vector length 2) in degrees around \verb{[lat,lon]} that will be searched. If discovered coordinates are outside \code{radius[1]}, a warning is issued, if they are outside \code{radius[2]}, an error is issued.}
}
\value{
A data.frame containing the datetime and variable values
}
\description{
Gathers, imports and collates an ERA5 variable from different netcdf files.
Typically ERA5 files are downloaded containing 1 variable at several coordinates for a certain period.
This function will search files that match a variable name string close to a given coordinate, read the data and assemble the variable from these files in chronological order. This tidies up ERA5 data downloads eg from era5cli program.
}
\details{
The location and variable name are kept in the attributes of the output.
}
\examples{
\dontrun{
ta <- gather_era5_var(namestrings = "2m_temperature", varid = "t2m",
datapath="../nc_input/era5_north_aral/",
lon=60.25, lat=46.5)
}
}
\author{
Tom Shatwell
ta <- gather_era5_var(namestrings = "2m_temperature", varid = "t2m",
}
