% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_phenology.R
\name{ice_phenology}
\alias{ice_phenology}
\title{Calculates ice phenology}
\usage{
ice_phenology(H_ice, dates, NH = TRUE)
}
\arguments{
\item{H_ice}{A vector indicating the presence of ice cover. Can be numerical or logical where \code{TRUE} or any number > \code{0} (e.g. ice thickness, indicates ice.
\code{FALSE} or \code{0} indicates no ice cover.}

\item{dates}{The dates corresponding to the ice observations in POSIX style and same length as \code{H_ice}.}

\item{NH}{Is the lake in the northern hemisphere? (logical)}
}
\value{
A \code{data.frame} containing ice stratification phenology
}
\description{
Calculates seasonal ice phenology including annual mean,
max and total duration of ice cover.
}
\details{
To be added later.
}
\examples{
\dontrun{
add later
}

}
\author{
Tom Shatwell
}
