% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bbe.R
\name{read_bbe}
\alias{read_bbe}
\title{Reads profile data from raw output file from a CTM CTD}
\usage{
read_bbe(
  filename,
  keep = c("depth", "sample_temperature", "Green_Algae", "Bluegreen", "Diatoms",
    "Cryptophyta", "Yellow_substances", "total_conc"),
  clean = TRUE,
  TZ = "UTC"
)
}
\arguments{
\item{filename}{Name of the file to read, including the file extension}

\item{keep}{names of the columns to retain in the output. Defaults to \code{"all"}}

\item{clean}{Should the profile be cleaned? (logical)}

\item{TZ}{time zone given to the output datetimes, defaults to \code{"UTC"} (recommended)}
}
\value{
A \code{data.frame} containing the profile data
}
\description{
Reads a data from an output file generated by a CTM CTD.
}
\details{
This function first collects information from the filename using \code{get_profile_info()},
including the sampling location from the beginning of the filename,
and the sonde number from the file extension.
It parses the column names from the file, and then reads the data.
It converts the date and time columns to a POSIX datetime with the timezone \code{TZ}.
The function finally drops any columns that are not named in \code{keep}.
The new column \code{pr_dt} is a unique timestamp for that profile, which is useful when compiling lots of profiles.
If \code{keep} contains names not found in the header, it will add a column of \code{NA}s for each unfound name.
If you choose to \code{clean} the profile, this removes any air values, values from the upward cast,
and the lowest 0.5 m of the profile to avoid sediment contact.
}
\examples{
\dontrun{
bbe_loc <- system.file("extdata", package="seefo")
bbe_name <- file.path(bbe_loc,  "profiles","20170801","YT1_20170801_1..FP2101")
bbe <- read_bbe(bbe_name)
head(bbe)
}

}
\seealso{
\code{\link{get_profile_info}},  \code{\link{read_ctm}}, \code{\link{clean_profile}}, \code{\link{batch_read}}
}
\author{
Tom Shatwell
}
